fix user retention tests
